ml-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
scikit-learn==1.4.2
numpy==1.26.4
kafka-python==2.0.2
pydantic==2.6.4
